pyproject.toml
[tool.poetry]
name = "strava-v2"
version = "0.1.0"
description = ""
authors = ["kennycdw <dewei274@gmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.10"
pandas = "^2.2.3"
fastapi = "^0.115.6"
python-dotenv = "^1.0.1"
asyncpg = "^0.30.0"
requests = "^2.32.3"
uvicorn = "^0.34.0"
python-telegram-bot = "^21.9"
loguru = "^0.7.3"
asyncio = "^3.4.3"
pyngrok = "^7.2.2"
nest-asyncio = "^1.6.0"
langchain = "^0.3.13"
langgraph = "^0.2.60"
google-generativeai = "^0.8.3"
langchain-google-genai = "^2.0.7"
tabulate = "^0.9.0"
pre-commit = "^4.1.0"
ruff = "^0.9.6"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
include = ["*.py"]
line-length = 150
target-version = "py310"